SLING-13170 Extract ManualScheduler to its own file
Move ManualScheduler from a private inner class in
JobManagerConfigurationTest to a standalone package-private class
for better readability.
